requirements.txt
pandas
streamlit
rapidfuzz
pytesseract
pillow
streamlit-lottie
